Initial commit of the create_pax script.
create_pax is a script to checksum and archive a directory tree.

You give it a directory tree, along with a path to put output files.
The output files are...

* A .pax.gz file, which is the archive (gzip-compressed).
* A .txt file, which is the catalog.

The catalog contains the following info for each item archived:

* The UNIX permissions.
* For files, the SHA-1 checksum.
* For files and symlinks, the filesize.
* The modification date.
* The file name.
* For symlinks, the target.

The catalog exists so that you can easily search for stuff, before you
go trawling through the entire archive.  Catalog entries are grouped
by directory; the directory's absolute path is placed at the top of
the group.

Note that because of the way Python buffers writes, the catalog file
will remain empty for some time after starting a run, even as the
archive file starts filling up.

Files are written in the pax format, which is the POSIX.1-2001-defined
format meant to "unify" tar and cpio standards.  I like it because...

* It supports long file names and link names.
* It stores names UTF-8-encoded.
* You can attach additional key/value metadata to entries (although I
  don't use that feature here).
* The `tar` command recognizes it.

It uses only stuff that comes with Python, and has been written for
(and tested with) Python 3.4.  It should work with newer Pythons, but
will _not_ work with Python 2.7.